add templates and dynamically load them. (#62)
Fix #8 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-8--petplace--hlxsites.hlx.page/

This PR adds the structure for templates, and provides the ability for
pages to specify which template should be used. Currently there are 3
templates:

* `article`: for information pages similar to
https://www.petplace.com/article/dogs/pet-care/dog-care/grooming-your-dog/the-ultimate-how-to-guide-for-grooming-a-dog-at-home/
* ` bio`: for breed pages similar to
https://www.petplace.com/article/breed/bernese-mountain-dogs/
* `category`: for pages with tiles, similar to
https://www.petplace.com/article/breed/

Pages can specify which template to use through their metadata, similar
to:

| Metadata | |
| --------- | --------- |
| Template | article |

Specifying this template metadata will cause the page to load the
`article.css` stylesheet, which will apply the general structure and
styles of that template.

The css files are currently empty, but I'm working on filling them out.
I wanted to get this PR merged early so that we could at least start
specifying the template metadata on the pages we're focusing on.