flesh out pc resolution for article template (#78)
Please always provide the [GitHub issue(s)](../issues) your PR is for,
as well as test URLs where your change can be observed (before and
after):

Fix #8 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-8-article-full--petplace--hlxsites.hlx.page/

This PR removes a couple of blocks, in favor of building the elements
inline (this cleans up the DOM). Adds styling for the largest resolution
layout of the article template.