Implement article navigation block (#92)
Fix #44 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-44--petplace--hlxsites.hlx.page/

Here is an example page with the navigation:
https://issue-44--petplace--hlxsites.hlx.page/article/cats/pet-care/best-halloween-themed-cat-names

Here is an example page _without_ the navigation:
https://issue-44--petplace--hlxsites.hlx.page/article/dogs/pet-care/dog-care/grooming-your-dog/the-ultimate-how-to-guide-for-grooming-a-dog-at-home

The algorithm for displaying next previous is:

* Get `/categories.json` and filter for all articles in the current
article's category.
* Sort the articles by the date they were created.
* Find the current article in the category list.
* The previous article will be the next older article in the list (it
will wrap around to the end if the current article is the first one)
* The next article will be the next newer article in the list (it will
wrap around to the beginning if the current article is the last one)

The article navigation won't be displayed at all if there are less than
3 articles in the category (i.e. there aren't enough for the previous,
current, and next).

In terms of layout, it uses a CSS grid to position all the items in the
block. As the window width grows, the space between the previous/next
card will also grow, up to the section maximum width.

The block changes quite a bit on mobile devices, so I'd recommend
switching device types to see the need for some of the extra divs.