feat: build responsive hero images (#108)
Fix #84 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-84-hero--petplace--hlxsites.hlx.page/

This change will use `createResponsiveImage` to make the hero image
responsive. It's based on the number of images appearing before the
first `h1` in the authored document. The logic is as follows:

* If there are 3 images before the header, the first will be used for
mobile, the second for tablet, and the third for desktop.
* When there are 2 images before the header, the first will be used for
both mobile and tablet, and the second for desktop.
* If there is only 1 image, it will be used for all screen sizes.

The change also incorporates `createOptimizedImage` to replace the hero
images with a rendition optimized for the screen width.

This page has responsive hero images:

-
https://issue-84-hero--petplace--hlxsites.hlx.page/authors/dr-debra-primovic-dvm

Additional URLs for testing:

-
https://issue-84-hero--petplace--hlxsites.hlx.page/article/dogs/pet-care/dog-care/grooming-your-dog/the-ultimate-how-to-guide-for-grooming-a-dog-at-home
-
https://issue-84-hero--petplace--hlxsites.hlx.page/article/breed/pembroke-welsh-corgi
- https://issue-84-hero--petplace--hlxsites.hlx.page/article/breed/