feat: add author info block (#114)
Fix #46

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-46--petplace--hlxsites.hlx.page/

Adds a new template (incomplete) for an "ask author" page. The very
basic layout for the template is in place - for now it's very similar to
the category info template.

The main task was to add a block for displaying an author's avatar and
name. The block is responsive, so it will change depending on screen
resolution. See newly added page:

-
https://issue-46--petplace--hlxsites.hlx.page/tags/ask-dr-debra-common-questions