feat: add support for hidden descriptive text (#145)
Fix #125 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-125--petplace--hlxsites.hlx.page/

As describe in the ticket, add new method that will create hidden
descriptive text for screen readers. As an example, see the "Traveling
with a Pet" slide's "Read More" button. The button itself still says
"Read More", but upon inspecting the button there will now be a span
whose class is `sr-only` and whose content is "about Traveling with a
Pet".