feat: add accessibility to slideshow (#140)
Fix #88

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-88--petplace--hlxsites.hlx.page/

This PR introduces a new method that will decorate an _existing_
slideshow block with attributes and functionality to meet accessibility
requirements. Some of the things provided are:

* Keyboard interaction support for slide selection tabs, if supported by
the slideshow block.
* Auto rotation capabilities, if desired by the block.
* Full accessibility-related attribute decoration for all elements of a
slideshow, including the slideshow's overall container, its tab
selection elements (if provided), and the slides themselves.
* The decorator doesn't do the work of actually visibly changing the
active slide, but it will decorate active/inactive slides appropriately,
and provide a custom event that consumers can use to do that work.