feat: add a11y "skip to" links (#147)
Fix #48 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After: https://issue-48--petplace--hlxsites.hlx.page/

This PR adds "skip to" buttons to the site. The buttons are hidden by
default, but are revealed when they're focused. They're the first
focusable elements on the page, so pressing the tab button will reveal
them. Interacting with the buttons will focus the corresponding location
in the page:

* "Skip to Content" will focus the `<main>` element.
* "Skip to Menu" will focus the first `li`'s `<a>` element in the page's
primary navigation.
* "Skip to Footer" will focus the `<footer>` element.

All of the "Skip to" buttons are in a containing `<nav>` element, which
is fixed, covers the entire content of the window, has a transparent
background, and a `z-index` of -1. When anything in the `nav` is
focused, it's brought to the front with a transparent black background;
this is what provides the slightly darker "overlay" of the entire page
whenever a "skip to" button is visible.

---------

Co-authored-by: Julien Ramboz <ramboz@users.noreply.github.com>