chore: try optimized font loading for smoother UX (#288)
Track if fonts have previously been loaded and cached by the browser so
we can speed up the loading on subsequent visits. If the browser cache
is not disabled, this should remove the flickering from the font swap
when you navigate across pages.

This is a variant of #286 that also introduces an optimized way to load
multiple requests for mobile vs. desktop so we don't impact the score in
a negative way. This ultimately helps improve LCP, TTI and SI scores
even further than the previous.

Fix #274 

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After:
  - https://issue-274c--petplace--hlxsites.hlx.page/
  - https://issue-274c--petplace--hlxsites.hlx.page/?martech=off