refactor: use placeholders for static strings to prepare for multi-language support (#365)
As part of #364, we need to extract all hardcoded strings so they can be
localized as needed. We also remove all the testing logic at the same
time since this isn't used anymore so we have a cleaner baseline for
future regional sites.

For this, a new `/placeholders.xslx` sheet has been added to Sharepoint,
and the site now pulls the static strings from there. We'll be able to
transparently replace the `placeholders.xslx` sheet for each locale down
the road so we have properly localized strings across the website.

We also introduce a new `getPlaceholder(key, options)` method in the
`scripts.js` so we can easily pull strings and quickly inject variables
in them (uses a `{{variableName}}` syntax)

Test URLs:
- Before: https://main--petplace--hlxsites.hlx.page/
- After:
  - https://placeholders--petplace--hlxsites.hlx.live/
  - https://placeholders--petplace--hlxsites.hlx.live/?martech=off